+cleanup
git-svn-id: https://svn.code.sf.net/p/remotefs/code/trunk@923 7764f161-3e89-4957-9ae6-8e1f4830ae4a

+++ src/nss_server.c
@@ -39,11 +39,11 @@ static char* nss_socket_name(struct rfs_instance *instance)
 
 	if (snprintf(name, 
 		FILENAME_MAX, 
-		"%srfs_nss-%d-%s.%d", 
+		"%srfs_nss-%d-%s.%lu", 
 		NSS_SOCKETS_DIR, 
 		uid, 
 		instance->config.host, 
-		pid) >= FILENAME_MAX)
+		(long unsigned)pid) >= FILENAME_MAX)
 	{
 		return NULL;
 	}
